fix: resolve error in current_user method
原本的程式當我現在是登入狀態卻將自己帳號刪除, 會因為在資料庫中找不到目前
session 紀錄的使用者 (因為還沒登出, 帳號卻被刪除) 而產生錯誤

現在將 User.find_by_id! 改成 User.find_by_id (找不到時直接回傳 nil)

並且在刪除使用者的時候先檢查該使用者是不是當前登入帳號,
如果是就會在刪除後也一併登出

+++ app/controllers/users_controller.rb
@@ -42,10 +42,19 @@ class UsersController < ApplicationController
     end
 
     def destroy
-        if @user.destroy
-            redirect_to users_url, notice: "User was successfully destroyed." # TODO i18n
+        if current_user == @user
+            if @user.destroy
+                reset_session # 如果把自己刪除, 就要登出
+                redirect_to login_path, notice: "Your account has been successfully deleted. You have been logged out." # TODO i18n
+            else
+                redirect_to users_url, alert: "Failed to destroy the user." # TODO i18n
+            end
         else
-            redirect_to users_url, alert: "Failed to destroy the user." # TODO i18n
+            if @user.destroy
+                redirect_to users_url, notice: "User was successfully destroyed." # TODO i18n
+            else
+                redirect_to users_url, alert: "Failed to destroy the user." # TODO i18n
+            end    
         end
     end
